first commit, of tutorial files
i have rewritten the tutorial code over and changed the names of locations and characters, my code is now functioning and i will next change the code more to become and more interesting and smooth running playable game that represents the game i planned.

Today i corrected some errors and make the currant location show
i also added a intro and instructions for players.

i have now added more locations and enemys to make the game more interesting so that it goes for longer and is more fun